Data member evaluation and compilation in progress.

+++ Assembler/Spect.Net.TestParser/SyntaxTree/Expressions/AddOperationNode.cs
@@ -54,7 +54,7 @@ namespace Spect.Net.TestParser.SyntaxTree.Expressions
             }
 
             // --- Numeric operands
-            return new ExpressionValue((ushort)(leftValue.AsNumber() + rightValue.AsNumber()));
+            return new ExpressionValue(leftValue.AsNumber() + rightValue.AsNumber());
         }
 
         public AddOperationNode(ParserRuleContext context) : base(context)
