+++ Assembler/Spect.Net.TestParser/SyntaxTree/Expressions/BitwiseOrOperationNode.cs
@@ -57,7 +57,7 @@ namespace Spect.Net.TestParser.SyntaxTree.Expressions
             }
 
             // --- Numeric operands
-            return new ExpressionValue((ushort)(leftValue.AsNumber() | rightValue.AsNumber()));
+            return new ExpressionValue(leftValue.AsNumber() | rightValue.AsNumber());
         }
 
         public BitwiseOrOperationNode(ParserRuleContext context) : base(context)
