+++ Assembler/Spect.Net.TestParser/SyntaxTree/Expressions/DivideOperationNode.cs
@@ -42,7 +42,7 @@ namespace Spect.Net.TestParser.SyntaxTree.Expressions
                 return ExpressionValue.Error;
             }
 
-            return new ExpressionValue((ushort)(leftValue.AsNumber() / rightValue.AsNumber()));
+            return new ExpressionValue(leftValue.AsNumber() / rightValue.AsNumber());
         }
 
         public DivideOperationNode(ParserRuleContext context) : base(context)
