+++ Assembler/Spect.Net.TestParser/SyntaxTree/Expressions/MultiplyOperationNode.cs
@@ -36,7 +36,7 @@ namespace Spect.Net.TestParser.SyntaxTree.Expressions
             }
 
             // --- Numeric operands
-            return new ExpressionValue((ushort)(leftValue.AsNumber() * rightValue.AsNumber()));
+            return new ExpressionValue(leftValue.AsNumber() * rightValue.AsNumber());
         }
 
         public MultiplyOperationNode(ParserRuleContext context) : base(context)
